refactor: moved generateLeagueID into the league use case
